+++ tests/Widgets/Label.cpp
@@ -247,7 +247,7 @@ TEST_CASE("[Label]")
             REQUIRE_NOTHROW(renderer->setProperty("TextOutlineThickness", 2));
             REQUIRE_NOTHROW(renderer->setProperty("Borders", tgui::Borders{1, 2, 3, 4}));
             REQUIRE_NOTHROW(renderer->setProperty("Padding", tgui::Borders{5, 6, 7, 8}));
-            REQUIRE_NOTHROW(renderer->setProperty("TextStyle", tgui::TextStyle{tgui::TextStyle::Bold | tgui::TextStyle::Italic}));
+            REQUIRE_NOTHROW(renderer->setProperty("TextStyle", tgui::TextStyles{tgui::TextStyle::Bold | tgui::TextStyle::Italic}));
             REQUIRE_NOTHROW(renderer->setProperty("TextureBackground", textureBackground));
             REQUIRE_NOTHROW(renderer->setProperty("Scrollbar", scrollbarRenderer.getData()));
             REQUIRE_NOTHROW(renderer->setProperty("ScrollbarWidth", 15));