+++ changelog.md
@@ -33,6 +33,7 @@ TGUI 0.9-dev
 - Changed default value of ScrollbarPolicy in Label to Automatic
 - Renamed getMenuList in MenuBar to getMenus
 - Renamed mouseOnWidget to isMouseOnWidget
+- Renamed TextStyle class to TextStyles and TextStyles::Style enum to TextStyle
 - Default scroll amount in ScrollablePanel now depends on global text size
 - Gui::getFont now returns the global font if no font was set in the Gui
 - Container now translates the widget position before calling draw function